ground accelerometer

+++ app/src/main/java/com/example/myapplication/Velocity.java
@@ -1,26 +1,26 @@
 package com.example.myapplication;
 
 public class Velocity {
-    private int x, y;
+    private float x, y;
 
-    public Velocity(int x, int y) {
+    public Velocity(float x, float y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
